Fix ESP8266 build warnings

+++ simple_ctrl.c
@@ -613,8 +613,8 @@ static void simple_ctrl_body_handle(void)
 								  (buffer[3] << 8) |
 								  (buffer[4] << 16) |
 								  (buffer[5] << 24);
-						ESP_LOGI(TAG, "(%d) load_type:%02x, crypto_type:%02x, load_len:%lu", fds[index],
-							 handle.load_type, handle.crypto_type, handle.load_len);
+						ESP_LOGI(TAG, "(%d) load_type:%02x, crypto_type:%02x, load_len:%u", fds[index],
+							 handle.load_type, handle.crypto_type, (unsigned int)handle.load_len);
 
 						count = 0;
 						while (count < handle.load_len) {
@@ -626,10 +626,11 @@ static void simple_ctrl_body_handle(void)
 								ESP_LOGI(TAG, "Read exception or timeout, disconnected (%d)", fds[index]);
 								goto closefd;
 							}
-							ESP_LOGD(TAG, "(%d) Expect: %lu; Result: %d", fds[index], size, ret);
+							ESP_LOGD(TAG, "(%d) Expect: %u; Result: %d", fds[index], (unsigned int)size, ret);
 
 							count += ret;
-							ESP_LOGD(TAG, "(%d) Handle [%lu/%lu]", fds[index], count, handle.load_len);
+							ESP_LOGD(TAG, "(%d) Handle [%u/%u]", fds[index],
+								 (unsigned int)count, (unsigned int)handle.load_len);
 
 							/* Handle data */
 							ret = simple_ctrl_handle_pad(&handle, buffer, ret, sizeof(buffer));
